Refactored number saver so it works off a generic implementation
One that lets subclasses decide which var types they can support. This should cut down on boilerplate code we'd otherwise make when working on other var encoders.
