+++ Assets/Fungus/Scripts/Save/Encoders/NumericVarSaveEncoder.cs
@@ -1,51 +1,19 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
+﻿using System.Collections.Generic;
+using Type = System.Type;
 
 namespace Fungus
 {
     /// <summary>
     /// This handles encoding numeric variables into Save Data.
     /// </summary>
-    public class NumericVarSaveEncoder: MonoBehaviour
+    public class NumericVarSaveEncoder: VarSaveEncoder
     {
-        protected static System.Type[] numericTypes = new System.Type[]
+        protected override IList<Type> SupportedTypes { get; } = new List<Type>
         {
             typeof(IntegerVariable),
             typeof(FloatVariable),
             typeof(BooleanVariable) // Yes, we count bools as numerics
         };
 
-        public virtual StringPair Encode(Variable varToEncode)
-        {
-            // I know you may be looking at this implementation and thinking
-            // "why not just have this class encode vars in general?"
-            // The reason is, I believe it's best to have encoders and decoders
-            // hot-swappable so that users can customize the encoding and decoding
-            // processes. After all, what if they decide that they want certain variable
-            // types to be encoded in ways different than the default for the sake of
-            // security or something?
-            StringPair encodedVar = new StringPair();
-            encodedVar.key = varToEncode.name;
-            encodedVar.val = varToEncode.GetValue().ToString();
-
-            return encodedVar;
-        }
-
-        public virtual IList<StringPair> Encode(IList<Variable> varsToEncode)
-        {
-            IList<StringPair> results = new StringPair[varsToEncode.Count];
-
-            for (int i = 0; i < varsToEncode.Count; i++)
-            {
-                var currentVar = varsToEncode[i];
-                var encodedVar = Encode(currentVar);
-                results[i] = encodedVar;
-            }
-
-            return results;
-        }
-
-        
     }
 }
