+++ Assets/Fungus/Scripts/Save/Encoders/VarSaveEncoder.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8f35c6739066c3b4990a899b557661a0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
